Move spf<String> validation to its own file
